requirements.txt
asgiref==3.5.2
certifi==2022.9.24
colorama==0.4.5
distlib==0.3.6
dj-database-url==0.5.0
Django==4.1
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.8.0
gunicorn==20.1.0
pip-install==1.3.5
platformdirs==2.5.2
psycopg2==2.9.1
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dateutil==2.8.2
python-snippets==0.0.3
pytz==2022.2.1
sib-api-v3-sdk==7.5.0
six==1.16.0
sqlparse==0.4.1
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.3
urllib3==1.26.12
virtualenv==20.16.5
whitenoise==5.2.0
django-ckeditor
